requirements.txt
accelerate==1.1.1
bitsandbytes==0.44.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
cssselect==1.2.0
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.4
lxml==5.3.0
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pfzy==0.3.4
prompt_toolkit==3.0.48
psutil==6.1.0
PyYAML==6.0.2
readability-lxml==0.8.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
sgmllib3k==1.0.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
